requirements.txt
appdirs==1.4.4
args==0.1.0
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
clint==0.5.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Django==3.0.5
django-phonenumber-field==5.0.0
gunicorn==20.0.4
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
isort==4.3.21
joblib==0.15.1
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
mysql-connector-python==8.0.20
networkx==2.4
packaging==20.3
pafy==0.5.5
phonenumbers==8.12.9
pooch==1.1.0
protobuf==3.11.3
pycodestyle==2.5.0
pygubu==0.9.8.6
pyimage==0.1.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytube3==9.6.4
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
scipy==1.4.1
six==1.13.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
threadpoolctl==2.0.0
tifffile==2020.5.11
tqdm==4.46.0
typing-extensions==3.7.4.2
urllib3==1.25.9
whitenoise==5.2.0
wrapt==1.11.2
youtube-dl==2020.5.3
